services/ml_service/requirements.txt
fastapi
uvicorn
pydantic>=2
mlflow
scikit-learn==1.4.2
pandas
pyyaml
python-dotenv
catboost
numpy
prometheus-fastapi-instrumentator

#for model from ./models/requirements.txt
mlflow==2.7.1
autofeat==2.1.2
babel==2.17.0
cloudpickle==2.2.1
packaging==23.2
psutil==7.0.0
pyyaml==6.0.2
scipy==1.11.3

#prometheus
prometheus-client>=0.20.0